ref: Split out project actor into separate file
