+++ relay-server/src/actors/mod.rs
@@ -45,6 +45,7 @@ pub mod healthcheck;
 pub mod keys;
 pub mod outcome;
 pub mod project;
+pub mod project_cache;
 pub mod project_keys;
 pub mod server;
 pub mod upstream;
